refactor : implement 계층으로 이동 (#809)
